+++ src/components/Thoughts.js
@@ -20,7 +20,6 @@ import Post from "./Post.js";
 import FocusView from "./FocusView.js";
 import StackView from "./StackView.js";
 
-
 export default function Thoughts({ viewStyle, shouldBlurAll }) {
 	// the displayMode gets set to $TESTINGMODE after every subreddit change.
 	const match = useRouteMatch("/:subreddit");
@@ -78,21 +77,11 @@ export default function Thoughts({ viewStyle, shouldBlurAll }) {
 				// but now it makes a new request to get listings each time we enter the wrong sub
 				history.goBack();
 			});
-
 	};
 	React.useEffect(() => {
 		// TODO: USE PERSISTENCE LIBRARY.
 		// to avoid fetching all listings of a subreddit when the user only intends to view one. SINGLE PAGE LOAD
-		if (!match.isExact) {
-			return null;
-		}
-
-		// if(postsData?.length > 2) {
-		// 	return null;
-		// }
-		// const url = "https://www.reddit.com/r/Showerthoughts/top/?t=month";
-		// by default .json at the end pulls the hot listings
-		// is this try-catch useless lol.
+		if (!match.isExact) return null;
 		try {
 			// how do i load the top listings !!??
 			loadListings(subreddit);
@@ -100,27 +89,10 @@ export default function Thoughts({ viewStyle, shouldBlurAll }) {
 			alert("useEffect failed");
 			console.log(e);
 		}
-		return () => {
-			// idont think this function get called ever now.
-			// too bad i would have put setHaveListings(false) in here.
-			// ? wait i see this console.log so it means that this component gets unMounted on every sub change ?
-			// ? and also on Focus post View wow why.
-			// ? ok so its because of react router. https://stackoverflow.com/questions/33431319/prevent-component-be-unmounted-with-react-router/38477462, https://stackoverflow.com/questions/45917133/react-router-never-unmount-a-component-on-a-route-once-mounted-even-if-route-c#:~:text=20-,React%2Drouter%3A%20never%20unmount%20a%20component%20on%20a%20route%20once,mounted%2C%20even%20if%20route%20change&text=Where%20basically%20each%20component%20is%20mounted%2Funmounted%20on%20route%20change.
-			// ? is it tho idk shit.
-			// setPostsData([]);
-			// EITHER DUE TO SUB CHANGE OR ISEXACT CHANGE SO IF ITS AN ISEXACT CHANGE WE SHOULD NOT CLEAR
-			// THE LISTINGS. IF THE URL IS NOT EXACT CURRENTLY exact then its focus mode.
-			// wait is this memoised ????
-			// if (isExact) { // ! added for memoisation concerns
-			// 	setPostsData([]);
-			// }
-			// alert("cleaning posts")
-			// will we still have to deal with the grid-overlapping if the clearing data works ?
-			// wait what sub shouldnt have to change when we move the FocusView check react-router-dom docs for this
-		};
-	// ok so i got it putting isExact in the dependency array was the fix to listings not being loaded
-	// after single post load -> history.goBack(), but now it 
-	}, [subreddit, isExact]);
+		return () => {};
+		// ok so i got it putting isExact in the dependency array was the fix to listings not being loaded
+		// after single post load -> history.goBack(), but now it
+	}, [subreddit]);
 
 	const findComment = (link) =>
 		comments.filter(
@@ -218,7 +190,18 @@ export default function Thoughts({ viewStyle, shouldBlurAll }) {
 			{/*should we add a powerbar here to control the view styles etc ?? */}
 			<Switch>
 				<Route exact path="/:subreddit">
-					{dataReceived && <StackView {...{postsData, loadMorePosts, expandView, shouldBlurAll, postsSeen, lastSeen}}></StackView>}
+					{dataReceived && (
+						<StackView
+							{...{
+								postsData,
+								loadMorePosts,
+								expandView,
+								shouldBlurAll,
+								postsSeen,
+								lastSeen
+							}}
+						></StackView>
+					)}
 				</Route>
 				{/*alternatively use useRouteMatch.url or what idk rn */}
 				<Route path={`/${subreddit}/:permalink`}>